Comment fix & double rather than float value

+++ victorious/victorious/UI/Content View/RealtimeComments/VRealtimeCommentsViewModel.m
@@ -141,7 +141,7 @@
 
 - (CGFloat)percentThroughMediaForCurrentRealTimeComment
 {
-    return !isnan(self.currentComment.realtime.floatValue / CMTimeGetSeconds(self.totalTime)) ? (self.currentComment.realtime.floatValue / CMTimeGetSeconds(self.totalTime)) : 0.0f;
+    return !isnan(self.currentComment.realtime.doubleValue / CMTimeGetSeconds(self.totalTime)) ? (self.currentComment.realtime.doubleValue / CMTimeGetSeconds(self.totalTime)) : 0.0f;
 }
 
 @end
